employee_microservice/employee_microservice/requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.16
djangorestframework==3.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.1.0
pika==0.12.0
pytz==2018.7
requests==2.20.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
